+++ pyproject.toml
@@ -36,7 +36,7 @@ classifiers = [
     "Typing :: Typed",
 ]
 dependencies = [
-    "ndv[pygfx]>=0.2.2",
+    "ndv[pygfx] @ git+https://github.com/tlambert03/ndv.git@streaming",
     "pydantic-settings>=2.7.1",
     "pymmcore-plus[cli]>=0.13.2",
     "pymmcore-widgets>=0.9.0",
@@ -44,7 +44,7 @@ dependencies = [
     "pyright>=1.1.393",
     "pyyaml>=6.0.2",
     "qtconsole>=5.6.1",
-    "rich",                         # also in pymmcore-plus[cli] above
+    "rich",                                                             # also in pymmcore-plus[cli] above
     "superqt[cmap,iconify]>=0.7.1",
     "tifffile>=2024.12.12",
     "tqdm>=4.67.1",
@@ -96,6 +96,8 @@ repository = "https://github.com/tlambert03/pymmcore-gui"
 line-length = 88
 target-version = "py310"
 src = ["src"]
+fix = true
+unsafe-fixes = true
 
 # https://docs.astral.sh/ruff/rules
 [tool.ruff.lint]
