+++ src/pymmcore_gui/_main_window.py
@@ -24,8 +24,8 @@ from pymmcore_gui.actions.widget_actions import WidgetActionInfo
 from ._ndv_viewers import NDVViewersManager
 from .actions import CoreAction, WidgetAction
 from .actions._action_info import ActionKey
-from .widgets._preview_viewer import LivePreview
 from .widgets._toolbars import OCToolBar, ShuttersToolbar
+from .widgets.image_preview._ndv_preview import NDVPreview
 
 if TYPE_CHECKING:
     from collections.abc import Callable, Mapping
@@ -127,7 +127,7 @@ class MicroManagerGUI(QMainWindow):
         # get global CMMCorePlus instance
         self._mmc = mmc = mmcore or CMMCorePlus.instance()
 
-        self._img_preview = LivePreview(self, mmcore=self._mmc)
+        self._img_preview = NDVPreview(self, mmcore=self._mmc)
         self._viewers_manager = NDVViewersManager(self, self._mmc)
 
         # MENUS ====================================
@@ -170,7 +170,7 @@ class MicroManagerGUI(QMainWindow):
         self.setCentralWidget(central_wdg)
 
         layout = QVBoxLayout(central_wdg)
-        layout.addWidget(self._img_preview.viewer.widget())
+        layout.addWidget(self._img_preview)
 
     @property
     def mmcore(self) -> CMMCorePlus:
